ref: Moved all closure related custom functions to its own file
