+++ CMakeLists.txt
@@ -36,6 +36,8 @@ add_library(Module SHARED
         Environment/Libraries/Filesystem.hpp
         ClosureManager.cpp
         ClosureManager.hpp
+        Environment/Libraries/Closures.cpp
+        Environment/Libraries/Closures.hpp
 )
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/Dependencies")
